reorg metrics module
